pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clubfridge-kasse"
version = "0.1.0"
description = "Clubfridge Digitale Kasse (Raspberry Pi)"
requires-python = ">=3.11"
dependencies = [
    "kivy>=2.3.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.36",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
# Nur auf dem Raspberry Pi / Linux installieren
pi = [
    "evdev>=1.6.0",
    "RPi.GPIO>=0.7.1",
]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
